Separated the mix source code
