+++ Source/User/Interface/UIComponents.cpp
@@ -22,235 +22,15 @@ along with Goblin Camp. If not, see <http://www.gnu.org/licenses/>.*/
 #include <boost/algorithm/string.hpp>
 
 #include "Goblin/User/Interface/UIComponents.hpp"
-#include "Goblin/User/Interface/Label.hpp"
 #include "Goblin/User/Interface/Dialog.hpp"
-#include "Goblin/User/Interface/ScrollPanel.hpp"
-#include "Goblin/User/Interface/Button.hpp"
 #include "UI.hpp"
 #include "Goblin/User/Interface/Grid.hpp"
-#include "Goblin/User/Interface/Spinner.hpp"
 #include "Goblin/User/Interface/Frame.hpp"
-#include "Goblin/User/Interface/TextBox.hpp"
 #include <Goblin/Config/WindowConfig.hpp>
 
 using namespace Goblin;
 
-void Label::Draw(int x, int y, TCODConsole* console)
-{
-	console->setAlignment(align);
-	console->setDefaultForeground(TCODColor::white);
-	console->print(x + _x, y + _y, text.c_str());
-}
 
-void LiveLabel::Draw(int x, int y, TCODConsole* console)
-{
-	console->setAlignment(align);
-	console->setDefaultForeground(TCODColor::white);
-	console->print(x + _x, y + _y, text().c_str());
-}
-
-void Button::Draw(int x, int y, TCODConsole *console) {
-	console->setBackgroundFlag(TCOD_BKGND_SET);
-	if(selected) {
-		console->setDefaultForeground(TCODColor::black);
-		console->setDefaultBackground(TCODColor::white);
-	} else {
-		console->setDefaultForeground(TCODColor::white);
-		console->setDefaultBackground(TCODColor::black);
-	}
-	console->setAlignment(TCOD_CENTER);
-	console->printFrame(x + _x, y + _y, width, 3);
-	console->print(x + _x + width/2, y + _y + 1, text.c_str());
-	console->setDefaultForeground(TCODColor::white);
-	console->setDefaultBackground(TCODColor::black);
-}
-
-void LiveButton::Draw(int x, int y, TCODConsole *console) {
-	Button::Draw(x, y, console);
-	if(selected) {
-		console->setDefaultForeground(TCODColor::black);
-		console->setDefaultBackground(TCODColor::white);
-	} else {
-		console->setDefaultForeground(TCODColor::white);
-		console->setDefaultBackground(TCODColor::black);
-	}
-	console->print(x + _x + width/2, y + _y + 1, textFunc().c_str());
-	console->setDefaultForeground(TCODColor::white);
-	console->setDefaultBackground(TCODColor::black);
-}
-
-MenuResult Button::Update(int x, int y, bool clicked, TCOD_key_t key) {
-	if(shortcut && (key.c == shortcut || key.vk == shortcut)) {
-		if (callback) {
-			callback();
-		}
-		return (MenuResult) ((dismiss ? DISMISS : 0) | KEYRESPOND);
-	}
-	if(x >= _x && x < _x + width && y >= _y && y < _y + 3) {
-		selected = true;
-		if(clicked && callback) {
-			callback();
-		}
-		return (MenuResult) (((clicked && dismiss) ? DISMISS : 0) | MENUHIT);
-	} else {
-		selected = false;
-		return NOMENUHIT;
-	}
-
-}
-
-void ToggleButton::Draw(int x, int y, TCODConsole *console) {
-	console->setBackgroundFlag(TCOD_BKGND_SET);
-	if(selected) {
-		console->setDefaultForeground(TCODColor::black);
-		console->setDefaultBackground(TCODColor::white);
-	} else {
-		console->setDefaultForeground(TCODColor::white);
-		console->setDefaultBackground(isOn() ? TCODColor::blue : TCODColor::black);
-	}
-	console->setAlignment(TCOD_CENTER);
-	console->printFrame(x + _x, y + _y, width, 3);
-	console->print(x + _x + width/2, y + _y + 1, text.c_str());
-	console->setDefaultForeground(TCODColor::white);
-	console->setDefaultBackground(TCODColor::black);
-}
-
-inline int numDigits(int num) {
-	int tmp = num;
-	int digits = 1;
-	if (tmp < 0) {
-		digits++;
-		tmp = -tmp;
-	}
-	while(num >= 10) {
-		digits++;
-		num /= 10;
-	}
-	return digits;
-}
-
-void Spinner::Draw(int x, int y, TCODConsole *console) {
-	console->setAlignment(TCOD_CENTER);
-	int val = value ? *value : getter();
-	console->print(x + _x + width / 2, y + _y, "- %d +", val);
-}
-
-MenuResult Spinner::Update(int x, int y, bool clicked, TCOD_key_t key) {
-	if ((x >= _x && x < _x + width && y == _y) || (key.vk == TCODK_KPADD || key.vk == TCODK_KPSUB)) {
-		if (clicked || (key.vk == TCODK_KPADD || key.vk == TCODK_KPSUB)) {
-			int curr = value ? *value : getter();
-			int adj = UI::Inst()->ShiftPressed() ? 10 : 1;
-			int strWidth = 4 + numDigits(curr);
-			if(x == _x + width / 2 - strWidth / 2 || key.vk == TCODK_KPSUB) {
-				if(value) {
-					(*value) = std::max(min, curr - adj);
-				} else {
-					setter(std::max(min, curr - adj));
-				}
-			} else if(x == _x + width / 2 + (strWidth-1) / 2 || key.vk == TCODK_KPADD) {
-				if(value) {
-					(*value) = std::max(min, curr + adj);
-				} else {
-					setter(std::max(min, curr + adj));
-				}
-			}
-		}
-		return MENUHIT;
-	}
-	return NOMENUHIT;
-}
-
-void TextBox::Draw(int x, int y, TCODConsole *console) {
-	console->setAlignment(TCOD_CENTER);
-	console->setDefaultBackground(TCODColor::darkGrey);
-	console->rect(x + _x, y + _y, width, 1, true, TCOD_BKGND_SET);
-	console->setDefaultBackground(TCODColor::black);
-	if(value) {
-		console->print(x + _x + width / 2, y + _y, value->c_str());
-	} else {
-		console->print(x + _x + width / 2, y + _y, getter().c_str());
-	}
-}
-
-MenuResult TextBox::Update(int x, int y, bool clicked, TCOD_key_t key) {
-	std::string currValue;
-	if(value) {
-		currValue = *value;
-	} else {
-		currValue = getter();
-	}
-	if(key.vk == TCODK_BACKSPACE && currValue.size() > 0) {
-		if(value) {
-			value->erase(value->end() - 1);
-		} else {
-			currValue.erase(currValue.end() - 1);
-			setter(currValue);
-		}
-		return KEYRESPOND;
-	} else if(key.c >= ' ' && key.c <= '}' && key.c != '+' && key.c != '-') {
-		if ((signed int)currValue.size() < width) {
-			if(value) {
-				(*value) += key.c;
-			} else {
-				currValue += key.c;
-				setter(currValue);
-			}
-		}
-		return KEYRESPOND;
-	}
-	return NOMENUHIT;
-}
-
-void ScrollPanel::Draw(int x, int y, TCODConsole *console) {
-	if (scroll < 0) {
-		scroll = 0;
-	}
-	if (scroll + height - 2 > contents->TotalHeight()) {
-		scroll = std::max(0, contents->TotalHeight() - (height - 2));
-	}
-	scrollBar = (int)((height-3) * ((double)scroll / (double)std::max(1, contents->TotalHeight() - (height-2)))) + 2;
-	scrollBar = std::min(scrollBar, height - 4);
-
-	if(drawFrame) {
-		console->printFrame(x + _x, y + _y, width, height);
-	}
-
-	contents->Draw(x + _x + 1, y + _y + 1, scroll, width - 2, height - 2, console);
-	console->putChar(x + _x + width - 2, y + _y + 1, TCOD_CHAR_ARROW_N, TCOD_BKGND_SET);
-	console->putChar(x + _x + width - 2, y + _y + height - 2, TCOD_CHAR_ARROW_S, TCOD_BKGND_SET);
-	console->putChar(x + _x + width - 2, y + _y + scrollBar, 219, TCOD_BKGND_SET);
-}
-
-MenuResult ScrollPanel::Update(int x, int y, bool clicked, TCOD_key_t key) {
-	if (x >= _x && x < _x + width && y >= _y && y < _y + height) {
-		if (clicked) {
-			if (x == _x + width - 2) {
-				if (y == _y + 1) {
-					scroll -= step;
-				} else if (y == _y + height - 2) {
-					scroll += step;
-				} else if (y < _y + scrollBar) {
-					scroll -= height;
-				} else if (y > _y + scrollBar) {
-					scroll += height; 
-				}
-			} else {
-				contents->Update(x - _x - 1, y - _y - 1 + scroll, clicked, key);
-			}
-		} else {
-			contents->Update(x - _x - 1, y - _y - 1 + scroll, clicked, key);
-		}
-		return MENUHIT;
-	}
-	if (key.vk != TCODK_NONE) return contents->Update(x, y, clicked, key);
-	return NOMENUHIT;
-}
-
-void ScrollPanel::GetTooltip(int x, int y, Tooltip *tooltip) {
-	if(x >= _x + 1 && x < _x + width - 1 && y >= _y + 1 && y < _y + height - 1) {
-		contents->GetTooltip(x - _x - 1, y - _y - 1 + scroll, tooltip);
-	}
-}
 
 void Frame::Draw(int x, int y, TCODConsole *console) {
 	console->printFrame(x + _x, y + _y, width, height, true, TCOD_BKGND_SET, title.c_str());
